remove userListElement and just use a one-off
